+++ common/src/main/java/dev/lpsmods/canned/block/CanBlock.java
@@ -1,8 +1,6 @@
 package dev.lpsmods.canned.block;
 
 import com.mojang.serialization.MapCodec;
-import net.minecraft.client.renderer.RenderType;
-import net.minecraft.client.renderer.entity.layers.RenderLayer;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.BlockGetter;