+++ common/src/main/java/dev/lpsmods/canned/core/ModTags.java
@@ -1,8 +1,6 @@
 package dev.lpsmods.canned.core;
 
-import dev.lpsmods.canned.Constants;
 import net.minecraft.core.registries.Registries;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Block;
@@ -14,20 +12,20 @@ public class ModTags {
     public static class Blocks {
         public static final TagKey<Block> CANS = createTag("cans");
 
-        public static void init() {}
+        public static void bootstrap() {}
 
         private static TagKey<Block> createTag(String name) {
-            return TagKey.create(Registries.BLOCK, ResourceLocation.fromNamespaceAndPath(Constants.MOD_ID, name));
+            return TagKey.create(Registries.BLOCK, ModUtils.makeId(name));
         }
     }
 
     public static class Items {
         public static final TagKey<Item> CANS = createTag("cans");
 
-        public static void init() {}
+        public static void bootstrap() {}
 
         private static TagKey<Item> createTag(String name) {
-            return TagKey.create(Registries.ITEM, ResourceLocation.fromNamespaceAndPath(Constants.MOD_ID, name));
+            return TagKey.create(Registries.ITEM, ModUtils.makeId(name));
         }
     }
 }