+++ common/src/main/java/dev/lpsmods/canned/core/ModUtils.java
@@ -1,0 +1,10 @@
+package dev.lpsmods.canned.core;
+
+import dev.lpsmods.canned.Constants;
+import net.minecraft.resources.ResourceLocation;
+
+public class ModUtils {
+    public static ResourceLocation makeId(String path) {
+        return ResourceLocation.fromNamespaceAndPath(Constants.MOD_ID, path);
+    }
+}