+++ fabric/src/main/java/dev/lpsmods/canned/Canned.java
@@ -1,11 +1,13 @@
 package dev.lpsmods.canned;
 
+import com.mrcrayfish.framework.FrameworkSetup;
 import net.fabricmc.api.ModInitializer;
 
 public class Canned implements ModInitializer {
     
     @Override
     public void onInitialize() {
+        FrameworkSetup.run();
         Bootstrap.init();
     }
 }